*** empty log message ***
git-svn-id: file:///usr/local/svn/fsc2@5762 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/src/module_util.c
@@ -277,7 +277,7 @@ double experiment_time( void )
 	t_old.tv_sec = t_new.tv_sec;
 	t_old.tv_usec = t_new.tv_usec;
 
-	return EDL.experiment_time += ( double ) dsec + 1.e-6 * ( double ) dusec;
+	return EDL.experiment_time += dsec + 1.e-6 * dusec;
 }
